requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.7
langchain==0.0.350
langchain-community==0.0.1
langchain-chroma==0.1.0
langchain-openai==0.0.2
chromadb==0.4.18
twilio==8.12.0
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
boto3==1.34.0
pygame==2.5.2
numpy==1.24.3
tiktoken==0.5.2
